EomBot/requirements.txt
discord.py==2.3.2
gspread==5.12.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
pytz==2023.3
aiohttp==3.9.1
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0